fix(Player): 수정된줄 알았지만 수정되지 않았던 버그 수정
