+++ src/test/java/ladder/model/LadderGeneratorTest.java
